requirements.txt
asn1crypto==0.24.0
atomicwrites==1.3.0
attrs==19.1.0
beautifulsoup4==4.8.1
bs4==0.0.1
cachetools==3.0.0
caldav==0.5.0
certifi==2018.11.29
cffi==1.11.5
chardet==3.0.4
coverage==4.5.2
cryptography==2.4.2
Django==2.2.6
emoji==0.5.1
future==0.17.1
google-api-python-client==1.7.8
google-auth==1.6.1
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.2.0
httplib2==0.12.0
idna==2.8
importlib-metadata==0.23
lxml==4.2.5
more-itertools==7.2.0
mysql-connector==2.2.9
mysql-connector-python==8.0.15
nose==1.3.7
oauth2client==4.1.3
oauthlib==3.0.1
packaging==19.2
pluggy==0.13.0
protobuf==3.7.0
psycopg2==2.7.7
psycopg2-binary==2.7.7
py==1.8.0
pyasn1==0.4.4
pyasn1-modules==0.2.2
pycparser==2.19
pyparsing==2.4.2
PySocks==1.6.8
pytest==5.1.2
python-dateutil==1.5
python-telegram-bot==11.1.0
pytz==2019.3
Radicale==2.1.11
requests==2.21.0
requests-oauthlib==1.2.0
rsa==4.0
schedule==0.5.0
six==1.12.0
soupsieve==1.9.5
sqlparse==0.3.0
uritemplate==3.0.0
urllib3==1.24.1
vobject==0.9.6.1
wcwidth==0.1.7
zipp==0.6.0
